backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.32
psycopg[binary]==3.2.1
alembic==1.13.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic-settings==2.4.0
pydantic[email]==2.8.2
bcrypt==3.2.2